refactor(fetchers): extract biorxiv + common modules, drop src/utils.py
GREEN phase for the src/fetchers/ split. All 27 tests pass.

src/fetchers/common.py:
- _ALLOWED_HOSTS, _validate_url, get_api_response (HTTP + URL validator)
- _load_existing_ids, load_all_existing_ids, filter_new_rows, write_file
  (shared dedup/IO; gain `dedup_cols=(2, 3)` keyword — biorxiv default
  (DOI, Version) at indices (2, 3); arxiv will pass (3, 4) for
  (ID, Version) in PR B3)

src/fetchers/biorxiv.py:
- parse_biorxiv_json, needs_pagination, build_date_range,
  prune_existing_csvs (biorxiv-specific)

src/app.py:
- Imports updated:
    from fetchers.biorxiv import build_date_range, needs_pagination,
        parse_biorxiv_json, prune_existing_csvs
    from fetchers.common import filter_new_rows, get_api_response,
        load_all_existing_ids, write_file
- PYTHONPATH=${{ github.action_path }}/src in action.yaml continues to
  resolve `from fetchers.* import ...` at runtime.

Deleted:
- src/utils.py (content fully migrated)
- tests/test_utils.py (content fully redistributed to
  tests/fetchers/test_{biorxiv,common}.py)

CHANGELOG: entry under [Unreleased] → Changed describing the split.
